requirements.txt
# Python dependencies for the Mediplus project
# Run `pip install -r requirements.txt` to install

setuptools
Django>=4.0,<5.0
# add extra packages as needed, e.g. django-environ, djangorestframework, etc.
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymysql==1.1.0
python-dotenv==1.0.0
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
stripe==7.5.0
boto3==1.34.0
django-storages==1.14.2
django-crontab==0.7.1
psycopg2-binary==2.9.9
django-sslserver==0.22.0
django-debug-toolbar==3.2.4